+++ src/__tests__/ledger.test.ts
@@ -5,8 +5,9 @@ import indy from 'indy-sdk'
 
 import { Agent } from '../agent/Agent'
 import { DID_IDENTIFIER_REGEX, VERKEY_REGEX, isFullVerkey, isAbbreviatedVerkey } from '../utils/did'
+import { sleep } from '../utils/sleep'
 
-import { genesisPath, getBaseConfig, sleep } from './helpers'
+import { genesisPath, getBaseConfig } from './helpers'
 import testLogger from './logger'
 
 const faberConfig = getBaseConfig('Faber Ledger', { genesisPath })
