+++ src/transport/index.ts
@@ -2,3 +2,4 @@ export * from './InboundTransporter'
 export * from './OutboundTransporter'
 export * from './HttpOutboundTransporter'
 export * from './WsOutboundTransporter'
+export * from './PollingInboundTransporter'
