+++ src/utils/sleep.ts
@@ -1,0 +1,3 @@
+export function sleep(ms: number) {
+  return new Promise((res) => setTimeout(res, ms))
+}
