+++ tests/__tests__/e2e.test.ts
@@ -1,7 +1,5 @@
-import type { InboundTransporter } from '../../src'
-
-import { Agent, AriesFrameworkError, HttpOutboundTransporter } from '../../src'
-import { getBaseConfig, sleep, waitForBasicMessage } from '../../src/__tests__/helpers'
+import { Agent, HttpOutboundTransporter, PollingInboundTransporter } from '../../src'
+import { getBaseConfig, waitForBasicMessage } from '../../src/__tests__/helpers'
 import logger from '../../src/__tests__/logger'
 import { get } from '../http'
 
@@ -93,39 +91,3 @@ describe('with mediator', () => {
     expect(basicMessage.content).toBe(message)
   })
 })
-
-class PollingInboundTransporter implements InboundTransporter {
-  public stop: boolean
-
-  public constructor() {
-    this.stop = false
-  }
-  public async start(agent: Agent) {
-    await this.registerMediator(agent)
-  }
-
-  public async registerMediator(agent: Agent) {
-    const mediatorUrl = agent.getMediatorUrl()
-
-    if (!mediatorUrl) {
-      throw new AriesFrameworkError(
-        'Agent has no mediator URL. Make sure to provide the `mediatorUrl` in the agent config.'
-      )
-    }
-
-    const mediatorInvitationUrl = await get(`${mediatorUrl}/invitation`)
-    const { verkey: mediatorVerkey } = JSON.parse(await get(`${mediatorUrl}/`))
-    await agent.routing.provision({
-      verkey: mediatorVerkey,
-      invitationUrl: mediatorInvitationUrl,
-    })
-    this.pollDownloadMessages(agent)
-  }
-
-  private async pollDownloadMessages(agent: Agent) {
-    while (!this.stop) {
-      await agent.routing.downloadMessages()
-      await sleep(5000)
-    }
-  }
-}
